- Updating based on comments. - Added base class to keep things a little more DRY
