+++ tdrs-backend/tdpservice/search_indexes/admin/filter.py
@@ -30,7 +30,7 @@ class CreationDateFilter(SimpleListFilter):
 
     def queryset(self, request, queryset):
         """Sort queryset to show latest records."""
-        if self.value() is None and len(queryset):
+        if self.value() is None and queryset.exists():
             max_date = DataFile.objects.all().latest('created_at').created_at
             datafile = DataFile.objects.get(created_at=max_date)
             return queryset.filter(datafile=datafile)
