+++ tdrs-backend/tdpservice/search_indexes/documents/ssp.py
@@ -1,25 +1,14 @@
 """Elasticsearch document mappings for SSP submission models."""
 
-from django_elasticsearch_dsl import Document, fields
+from django_elasticsearch_dsl import Document
 from django_elasticsearch_dsl.registries import registry
 from ..models.ssp import SSP_M1, SSP_M2, SSP_M3
 from tdpservice.data_files.models import DataFile
+from .document_base import DocumentBase
 
 @registry.register_document
-class SSP_M1DataSubmissionDocument(Document):
+class SSP_M1DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed SSP M1 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -82,21 +71,8 @@ class SSP_M1DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class SSP_M2DataSubmissionDocument(Document):
+class SSP_M2DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed SSP M2 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -183,20 +159,8 @@ class SSP_M2DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class SSP_M3DataSubmissionDocument(Document):
+class SSP_M3DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed SSP M3 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
