+++ tdrs-backend/tdpservice/search_indexes/documents/tanf.py
@@ -1,25 +1,13 @@
 """Elasticsearch document mappings for TANF submission models."""
 
-from django_elasticsearch_dsl import Document, fields
+from django_elasticsearch_dsl import Document
 from django_elasticsearch_dsl.registries import registry
 from ..models.tanf import TANF_T1, TANF_T2, TANF_T3, TANF_T4, TANF_T5, TANF_T6, TANF_T7
-from tdpservice.data_files.models import DataFile
+from .document_base import DocumentBase
 
 @registry.register_document
-class TANF_T1DataSubmissionDocument(Document):
+class TANF_T1DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T1 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -84,20 +72,8 @@ class TANF_T1DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class TANF_T2DataSubmissionDocument(Document):
+class TANF_T2DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T2 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -186,20 +162,8 @@ class TANF_T2DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class TANF_T3DataSubmissionDocument(Document):
+class TANF_T3DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T3 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -240,20 +204,8 @@ class TANF_T3DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class TANF_T4DataSubmissionDocument(Document):
+class TANF_T4DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T4 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -287,20 +239,8 @@ class TANF_T4DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class TANF_T5DataSubmissionDocument(Document):
+class TANF_T5DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T5 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -350,20 +290,8 @@ class TANF_T5DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class TANF_T6DataSubmissionDocument(Document):
+class TANF_T6DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T6 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
@@ -402,20 +330,8 @@ class TANF_T6DataSubmissionDocument(Document):
 
 
 @registry.register_document
-class TANF_T7DataSubmissionDocument(Document):
+class TANF_T7DataSubmissionDocument(DocumentBase, Document):
     """Elastic search model mapping for a parsed TANF T7 data file."""
-
-    datafile = fields.ObjectField(properties={
-                      'pk': fields.IntegerField(),
-                      'created_at': fields.DateField(),
-                      'version': fields.IntegerField(),
-                      'quarter': fields.TextField()
-                  })
-
-    def get_instances_from_related(self, related_instance):
-        """Return correct instance."""
-        if isinstance(related_instance, DataFile):
-            return related_instance
 
     class Index:
         """ElasticSearch index generation settings."""
